refactor(content): move all db interactions to FriendsServerRepository
